Fix issue #39
This fixes issue #39 and allows Tensorboard's SummaryWriter.add_graph to work although there are some TracerWarnings.

`TracerWarning: Converting a tensor to a Python number might cause the trace to be incorrect. We can't record the data flow of Python values, so this value will be treated as a constant in the future. This means that the trace might not generalize to other inputs!`